Rarer Minimal Labels
Previously when the minimal labels was created it ran on the obvious PR event
types.  It turns out that running it on open was problematic because there was
a race between the open event and the labeling events and often resulted in a
PR being improperly marked as failing.  This change updates the created
pipelines to not fire on open, but maintains all the rest of the PR event type
triggers.

Signed-off-by: Ben Hale <bhale@vmware.com>