Remove GCS Filter Prefix
Previously when a GCS bucket was listed, a prefix that was the equivalent of
the parent directory of the interesting keys was added.  Unfortunately, if
that parent directory had version numbers in it, it would break the listing.
This change removes that prefix so that all viable candidates are listed.

Signed-off-by: Ben Hale <bhale@vmware.com>