HTTP Credentials
The addition of the new Aqua Security buildpack has created a requirement that
downloads of dependencies be able to provide credentials for basic auth.
Rather than creates a new pinpoint deep in the packaging behavior of the
buildpacks, we can build on the fact that Golang honors a user's .netrc file.

This change updates the credential contribution in the descriptor to handle
both docker credentials and HTTP credentials.  The new HTTP credentials are
written into the workflow's .netrc file so that when create-package (and
others) run, they will have the credentials necessary for basic auth to some
servers.

Signed-off-by: Ben Hale <bhale@vmware.com>