Crane in Create Builder
Previously, the Create Builder job ran generally properly, however the final
step, updating a release with the digest of the builder in a registry did not
work because crane had not been installed before retrieving the digest.  And
unfortunately because of bash's suppression in subshells it took a first real
release to detect the problem.

This change ensures that the create-builder job installs crane before it needs
to be used.

Signed-off-by: Ben Hale <bhale@vmware.com>