Corretto Version Number
Previously, the version numbers used for comparison by the Amazon Corretto
Dependency action took the "extended" portion of the version number and place
it in semver metadata.  Metadata isn't used in ordered comparison by semver
and therefore this information, which did have semantic meaning, was being
ignored.

This change moves that same information to the pre-release portion of semver
where it will be used for comparison.

Signed-off-by: Ben Hale <bhale@vmware.com>