AdoptOpenJDK Version Number
Previously, the version numbers used for comparison by the AdoptOpenJDK
Dependency action took the semver version number, as declared by the API
including a portion in semver metadata.  Metadata isn't used in ordered
comparison by semver and therefore this information, which did have semantic
meaning, was being ignored.

This change moves that same information to the pre-release portion of semver
where it will be used for comparison.

Signed-off-by: Ben Hale <bhale@vmware.com>