Handle No Sub-Buildpacks
Previously when writing release notes, the script would attempt to iterate
over any sub-buildpacks that were contained within this main buildpack.  In
the case where there were no sub-buildpacks, an error would be printed to the
console.  This error was harmless, but better to be correct.  This change only
iterates over the sub-buildpacks if at least one exists.

Signed-off-by: Ben Hale <bhale@vmware.com>