PR Identity
Previously when a PR was created the author for the create commit was the
"actor" in the action.  In actions based on schedules, this actor is the
author of the last commit; the commit that the action reads its workflow from.
This behavior lead to commits appearing to come from real users rather than
the robots that were creating the PRs.

This change updates the author of the created commits to always be the same
bot that is creating the PR.

Signed-off-by: Ben Hale <bhale@vmware.com>