Guard against writing empty dependabot.yml files
Dependabot requires updates to be an array and it will fail if omitted or empty. The dependabot.yml file will need to be manually deleted from repos where it is current empty or no longer needed (now that we no longer subscribe to GHA updates) because the pipeline builder does not delete files.

Signed-off-by: Emily Casey <ecasey@vmware.com>