Fixes version ordering for GraalVM dependencies
Previously we normalizing GraalVM versions into semver format before sorting. With the release of 21.0.0.2 this caused a bug where version 21.0.0 was overwriting version 21.0.0.2 in the canidate map, causing version 21.0.0.2 to be ignored. With this change we no longer convert to semver before sorting and instead use a simple alphanumeric sort.

This is a quick fix and might be fragile to changes such as the introduction of pre-releases releases for release candidates.

Signed-off-by: Emily Casey <ecasey@vmware.com>