Enable setting two dependency ids
During the update package dependency workflow, a dependency id is required. This is used to update both the buildpack.toml and package.toml. However, the id is not always the same in both files. By default, the workflow will use the DEPENDENCY variable for both. You can however also set one of BP_DEPENDENCY or PKG_DEPENDENCY. Do not set both! You need to set DEPENDENCY and this is where the workflow will obtain the new version (it looks at the image referenced by the id, like gcr.io/paketo-buildpacks/bellsoft-liberica). You can then set the variable for the value which differs from DEPENDENCY. Typically, this will be BP_DEPENDENCY as the package.toml file uses the full image name in it where buildpack.toml has the buildpack id name.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>