Change the way Java stackdriver version is being represented
The Google Stackdriver Java agnet is using a date based version. Previously, this action would take the date part and stuff that into the major version field, followed by the RC number as the minor version. This resulted in the buildpack bumping a major version for every release, unless it's manually corrected.

This PR changes the transformation of the date to a version such that the major version won't change nearly as often. It takes the year from the date and sets that to the major version, the month to the minor version and the patch version becomes the combination of the date and the RC number. Ex: 2021.11.100.

It's not perfect, but we need a version in that format and this will ensure that we're not bumping major versions every time a new release drops.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>