Adds flag to update the version
Previously, update-buildpack-image-id would only allow you to update the id. In some cases, the id and version need to change in sync. For example of you have buildpack abc v5.4.10, you may want to map that to buildpack xyz v1.0.0. This will now rewrite the metadata such that both the id and the version are updated. The current id and version are pulled from the metadata on the existing buildpack (arg #1) image.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>