Add separate CPE & PURL version
Previously, we used same version across the version, CPE & PURL. The version was then updated to be a 3-digit version, and the system is attempting to use this 3-digit version in the CPE & PURL too (because that's the default behavior). Unfortunately, we want the full four digit version in the CPE & PURL for proper matching against vulnerabilities.

This PR makes the action export the 4-digit version for the CPE & PURL and the 3-digit version for the version.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>