Add if guard to compute-version.sh
This script is being run to determine the version, it requires `$GITHUB_REF` to be set and non-empty. In some rare cases, Github Actions are not populating `$GITHUB_REF` and this causes the script to incorrectly detect the version. This makes the `$GITHUB_REF` required and fails fast if it is not.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>