Adds a Github buildpack loader
The draft-release job loads buildpack info from remote sources. Previously, this was loaded from images, but the images can be quite large at times so that is slow. This PR adds a buildpack loader that fetches buildpack metadata from Github.

In some cases, image names do not map directly to Github repositories so this PR also adds 'mapper' input parameters which are regular expressions that can be used to rewrite the image URI to a Github project. For example, `mapper_input_1=|foo|bar|` will map foo in the image URI to bar. The code will try both the original unmapped value, as well as the mapped value. If you set multiple mappers, they are all tried in order with the first match that results in a successful load of buildpack metadata being the one that gets used.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>