Adds compute-artifact-description action
This action will go through and compute a list of all the dependencies used by the buildpack. It is recursive, so if you point it to a composite buildpack, it'll pull all of the dependencies for every buildpack included in that composite buildpack and give one list, sorted and with duplicates removed.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>