Bump docker/login-action@v1 to v2
Reviewed changes upstream in the action and the major verison bump appears to be due to them increasing the minimum Node.js version required to run the action from 12 to 16. This was similar in other actions, causing a major bump but not changing config or functionality in the action itself.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>