Sort by Strings is not totally accurate
In a couple of the actions, we were sorting an array of string versions. String versions do not sort correctly though under some circumstances, like if you have minor/patch releases in the list that are both less than and greater than 10. You get a sort like 1, 10, 11, 2, 3, 4, etc.. not 1, 2, 3, 4, 10, 11. This can cause updates to be missed.

This PR rewrites the two impacted actions such that they are using Semver to sort and pick the latest order, which guarantees that we get the most recent release. This PR also makes some optimizations so that the version number is loaded from metadata instead of the JVM file, or that we only download and read the version from at least one JVM.

In addition, it removes support for Java 8 from GraalVM, which dropped Java 8 support.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>