Add Update Go Workflow and Update to Go 1.18
This workflow runs weekly and updates both the Go version and the Go modules. This will update not just the direct dependencies but also transitive dependencies, which is something Dependabot does not presently do.

The Go version is determined by Pipeline Builder. When the Pipeline Builder workflow runs and updates the pipeline, it will update the required Go version in this workflow. The next time this workflow runs, it'll then bump the Go version. This workflow can be directly triggered as well.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>