Add NewActionContributionWithNamespace and deprecate NewActionContribution
This PR deprecates 'NewActionContribution' in favor of 'NewActionContributionWithNamespace'. You may still use the older method, but it will result in workflows being generated and not having a namespace which is not recommended. When using the new method, all workflows will be generated under the 'pb-*.yml' namespace. This allows the pipeline-builder to also remove old workflows when it updates itself. It will delete everything under 'pb-*.yml' and then regenerate all the files. The namespace also allows you to skip workflows you've custom crafted or generated with other tools, so long as they do not prefix with 'pb-*.yml'.

You may customize the prefix by passing in a namespace string to 'NewActionContributionWithNamespace'. The default used by workflows created by pipeline-builder is 'pb'.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>