Fixes "Update Go" job for libcnb
In libcnb, there are multiple sets of go.mod|sum files. One for the top level library and one for the tools. This is a fairly common pattern.

This PR fixes detection so that it only looks at the top level go.mod|sum files. If there are go.mod|sum files in nested directories they are ignore. It is not clear what we should do in the case of go.mod|sum files that are nested. We could potentially generate multiple "Update Go" workflow files or perhaps generate one that updates multiple locations. That will be a feature enhancement request though and not addressed here.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>