Allow a version filter for Tomcat
This PR adds a version filter for the Tomcat action. This allows for a more fine grained selection of the version that the dependency will check. By default, it'll look for the latest version of a major branch. For example, `10.*.*`. You can use this to lock the version to the latest of a particular minor branch, like `10.0.*`.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>