Add version regex pattern for Maven dependency action
This commit adds a `INPUT_VERSION_REGEX` input which can be used to filter the selected version.

It defaults to selecting any three-digit version number. You can set it to `^3\.[\d]+\.[\d]+$` to select only 3.x or `^4\.[\d]+\.[\d]+$` to only select 4.x. The only requirement is that the version regex match the given version string. The first match is the version selected.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>