Fix two small bugs
First. This fixes the case where running in a paketo-community repo. It prevents the Docker credentials from being added to the test workflow multiple times.

Second. This fixes an issue where you try to run `octo` outside of Github Actions, like on your local machine. In this case, `$GITHUB_REPOSITORY` probably will not be set and so the seed string for the cron jitter will be just `Update Go ` which is static and always returns the same number. This will unnecessarily change the cron value used in the `Update Go` job. This change will fallback to looking at the directory paths to try and find the right name to use. It assumes the folder structure is `paketo-buildpacks/<some-buildpack>` like the way repos are structured on Github. This will not be perfect, but it a bit better.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>