Add support for merge queues
This PR adds a merge group event type which can be added to workflows as a trigger. It exposes the documented types, of which there is only one, and it adds filters for branch and path (these are not documented, but I see references to them in other projects and they are needed to make this useful). It also updates the tests workflow so that tests are run when a merge group event is detected. It does not add label checks or action workflows. Those will run on PR, which is probably sufficient. We just want to re-run the tests in the merge queue to validate.

This PR has not been tested because we cannot use the merge queue feature yet. EasyCLA is not compatible. This is an attempt to prepare but a follow up PR may be required. Consider this experimental for now.

Resolves #1036

Signed-off-by: Daniel Mikusa <dan@mikusa.com>