Set CGO_ENABLED=0 when we build
This forces the Go compile tools to produce a statically linked binary. In Go 1.20, the default changed to producing a binary that is dynamically linked against libc. We could use that, but it can create compatibility issues if the build system has a newer libc version than on the stacks we support.

Presently, that happens because the buildsystems are on Jammy and we have to support the Bionic stack for a bit longer.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>