Change `go get -u` to fetch minor/patch updates
Previously we were using `get get -u all`, which was great but started pulling in major version updates which of course break things. This PR switches to not pull in major version bumps, which is the default behavior of `go get -u` and it adds the `-t` flag so we also update test modules. I think this is a safer approach, and it allows us to manually update major version bumps which will usually require manual intervention anyway.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>