Fix Rust Updater
The Rust updater looks a Github to pick the latest release. It was using the Github client's `GetLatestRelease` method, however, it looks like the Rust project doesn't cut releases for patch fixes so we have missed some of these updates.

This fix will use the Github client to fetch all of the tags and pick the latest release from the tags. There are tags even for patch releases.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>