Fix arm64 update check issue
This PR adds another `jq` selector to the query which is running to fetch the old version from buildpack.toml. Previously, this would return two version numbers, one for each version, and that casued a failure in the pipeline. With the additional selector, we get one version, the version for that specific architecture.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>