Bump action versions (#1480)
* Bump action versions

Bump versions for:
- docker/login-action
- mheap/github-action-required-labels
- peter-evans/create-pull-request

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Adjust logic for calculating version (#1481)

In most cases, `GITHUB_REF` is set to something that starts with `refs/*`. There is a bug where it's not set and when that happens, we get a bad release. The version calculation happens incorrectly and it uses the git commit hash for the version, but that does not work for an actual release.

There are three scenarios when the version is calculated:

- When publishing a builder
- When publishing a buildpack
- When running tests for publishing a builder (i.e. create-package)

The first two happen on release, and should read the tag and parse out the semver number. The latter should not have a version so it will fallback to running git to obtain the git commit hash and use that for a version.

We were previously trying to detect the case where the variable was unset or empty, but didn't have the logic right. This PR switches things so that we look to make sure `GITHUB_REF` starts with `refs/*`. I'm hoping this will be more reliable.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

---------

Signed-off-by: Daniel Mikusa <dan@mikusa.com>