Control build.sh and .gitignore
This PR adds control of the build.sh and .gitignore files over to `octo`. When it runs, it will generate those files. The build.sh script comes from a template file which is static at the moment. This could be modified to be a go template or something more flexible if necessary in the future. The .gitignore file is generated from a list of strings. For now, this ends up being static, but we could add in logic to generate the list dynamically if needed in the future.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>