Strip patch version off Go version
The input version is two digits (major.minor), but for some buildpacks the version in the go.mod file has three digits. This didn't actually change the go.mod file, but it caused a mismatch if old/new versions, which caused the wrong title to be filed on PRs.

Stripping off the patch version of the go version read out of the go.mod file should enable it to match unless the major.minor is different.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>